requirements.txt
streamlit
langchain
langgraph
openai
psycopg2-binary
sqlalchemy
python-dotenv
pandas
loguru
tiktoken
sqlaclchemy